requirements.txt
click<8.0,>=7.0  # To pass Celery
Django==3.2.5
psycopg2-binary==2.9.1
djangorestframework==3.12.4
requests==2.25.1
celery[redis]==5.1.1
